back/pyproject.toml
[tool.poetry]
name = "back"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.11"
fastapi = "^0.109.2"
uvicorn = "^0.27.1"
sqlalchemy = "^2.0.27"
httpx = "^0.26.0"
alembic = "^1.13.1"
pydantic-settings = "^2.2.0"
asyncpg = "^0.29.0"
pyjwt = "^2.8.0"
bcrypt = "^4.1.2"
cryptography = "^42.0.5"


[tool.poetry.group.dev.dependencies]
pytest = "^8.0.2"
black = "^24.2.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
